conversor e modificasoes de login

+++ src/main/java/br/edu/utfpr/currency_converter/model/domain/Currency.java
@@ -19,5 +19,5 @@ public class Currency{
 
     @NonNull
     @Column(name = "value_relative_real")
-    private double valueRelativeReal;
+    private float valueRelativeReal;
 }
